### 2021-04-17 #3177 @whedon
[PRE REVIEW]: morse: an R-package in support of Environmental Risk Assessment
**Submitting author:** @sandrinecharles (<a href="http://orcid.org/0000-0003-4604-0166">Sandrine CHARLES</a>)
**Repository:** <a href="https://github.com/pveber/morse" target ="_blank">https://github.com/pveber/morse</a>
**Version:** 3.3.1
