### 2021-04-20 #3177 @sandrinecharles
> **Submitting author:** @sandrinecharles ([Sandrine CHARLES](http://orcid.org/0000-0003-4604-0166))
> **Repository:** https://github.com/sandrinecharles/morse_paper
> **Version:** 3.3.1
> **Editor:** @KristinaRiemer
> **Reviewer:** Pending
> **Managing EiC:** Kyle Niemeyer
>
> **⚠️ JOSS reduced service mode ⚠️**
>
> Due to the challenges of the COVID-19 pandemic, JOSS is currently operating in a "reduced service mode". You can read more about what that means in [our blog post](https://blog.joss.theoj.org/2020/05/reopening-joss).
>
> **Author instructions**
>
> Thanks for submitting your paper to JOSS @sandrinecharles. **Currently, there isn't an JOSS editor assigned** to your paper.
>
> The author's suggestion for the handling editor is @KristinaRiemer.
>
> @sandrinecharles if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
>
> **Editor instructions**
>
> The JOSS submission bot @whedon is here to help you find and assign reviewers and start the main review. To find out what @whedon can do for you type:
>
> ```
> @whedon commands
> ```

Dear colleague,
As reviewers I may suggest the following ones:
Andre GERGS, andre.gergs@bayer.com
Alessio IPPOLITO, Alessio.IPPOLITO@efsa.europa.eu
Arnaud CHAUMOT, arnaud.chaumot@inrae.fr

Else, within your own list, those persons could maybe be interested in reviewing our manuscript:
@BrandonEdwards
@ghislainv
@Volkerschmid
@BrandonEdwards

Best regards,
Sandrine Charles
