### 2021-04-20 #3177 @sandrinecharles
> ```
> Software report (experimental):
>
> github.com/AlDanial/cloc v 1.88  T=0.02 s (245.9 files/s, 36241.3 lines/s)
> -------------------------------------------------------------------------------
> Language                     files          blank        comment           code
> -------------------------------------------------------------------------------
> Lua                              2             32             68            255
> Markdown                         2             57              0            220
> TeX                              1              9              0             96
> -------------------------------------------------------------------------------
> SUM:                             5             98             68            571
> -------------------------------------------------------------------------------
>
>
> Statistical information for the repository 'ba90c03f9d5b76567bf6507a' was
> gathered on 2021/04/17.
> No commited files with the specified extensions were found.
> ```

@whedons could you detail what is expected from my side as actions regarding this post?
