### 2021-04-20 afa65a9
Add paper folder with all files included
