+++ paper/paper.bib
@@ -1,0 +1,100 @@
+@article{Charles2017,
+author = {Charles, Sandrine and Delignette-Muller, Marie Laure and Veber, Philippe and Delignette-Muller, Marie Laure},
+doi = {10.1007/s11356-017-9809-4},
+file = {:Users/scharles/Documents/Mendeley Desktop/2018/Charles, Veber, Delignette-Muller/Charles, Veber, Delignette-Muller - 2018 - Environmental Science and Pollution Research.pdf:pdf},
+isbn = {1135601798},
+issn = {0944-1344},
+journal = {Environmental Science and Pollution Research},
+keywords = {Standard bioassay data,Survival statistical analys,analysis,r software,reprotoxicity statistical analysis,ssd,standard bioassay data,survival statistical},
+pages = {11295--11302},
+publisher = {Environmental Science and Pollution Research},
+title = {{MOSAIC: a web-interface for statistical analyses in ecotoxicology}},
+url = {http://link.springer.com/10.1007/s11356-017-9809-4},
+volume = {25},
+year = {2018}
+}
+
+@Manual{morse2021,
+    title = {morse: Modelling Tools for Reproduction and Survival Data in
+Ecotoxicology},
+    author = {Virgile Baudrot and Sandrine Charles and Marie Laure Delignette-Muller and Wandrille Duchemin and Benoit Goussen and Nils Kehrein and Guillaume Kon-Kam-King and Christelle Lopes and Philippe Ruiz and Alexander Singer and Philippe Veber},
+    year = {2021},
+    note = {R package version 3.3.0},
+    url = {https://CRAN.R-project.org/package=morse},
+  }
+  
+@article{baudrot2018a,
+author = {Baudrot, Virgile and Preux, Sara and Ducrot, Virginie and Pav{\'{e}}, Alain and Charles, Sandrine},
+doi = {10.1021/acs.est.7b05464},
+file = {:Users/scharles/Documents/Mendeley Desktop/2018/Baudrot et al/Baudrot et al. - 2018 - Environmental Science {\&} Technology.pdf:pdf;:Users/scharles/Documents/Mendeley Desktop/2018/Baudrot et al/Baudrot et al. - 2018 - Environmental Science {\&} Technology(2).pdf:pdf},
+journal = {Environmental Science {\&} Technology},
+number = {3},
+pages = {1582--1590},
+title = {{New insights to compare and choose TKTD models for survival based on an inter-laboratory study for Lymnaea stagnalis exposed to Cd}},
+url = {http://dx.doi.org/10.1021/acs.est.7b05464},
+volume = {52},
+year = {2018}
+}
+
+@article{Baudrot2019,
+author = {Baudrot, Virgile and Charles, Sandrine},
+doi = {10.1038/s41598-019-47698-0},
+file = {:Users/scharles/Documents/Mendeley Desktop/2019/Baudrot, Charles/Baudrot, Charles - 2019 - Scientific Reports, natureresearch.pdf:pdf;:Users/scharles/Documents/Mendeley Desktop/2019/Baudrot, Charles/Baudrot, Charles - 2019 - Scientific Reports, natureresearch.pdf:pdf;:Users/scharles/Documents/Mendeley Desktop/2019/Baudrot, Charles/Baudrot, Charles - 2019 - Scientific Reports, natureresearch(2).pdf:pdf},
+journal = {Scientific Reports, natureresearch},
+keywords = {dose-response,environmental risk assessment,guts,lethal concentration,margin of safety,multiplication factor,survival models},
+pages = {11432},
+publisher = {Springer US},
[…]
