### 2021-04-20 #3177 @sandrinecharles
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1007/s11356-017-9809-4 is OK
> - 10.1021/acs.est.7b05464 is OK
> - 10.2903/j.efsa.2018.5377 is OK
> - 10.1002/ieam.4061 is OK
>
> MISSING DOIs
>
> - 10.1007/s10646-012-0917-0 may be a valid DOI for title: Toxicokinetic-toxicodynamic modelling of survival of Gammarus pulex in multiple pulse exposures to propiconazole: model assumptions, calibration data requirements and predictive power
>
> INVALID DOIs
>
> - http://dx.doi.org/10.1101/356469 is INVALID because of 'https://doi.org/' prefix
> ```

File paper.bib has been corrected. Thanks for your vigilance.
