### 2021-04-20 #3177 @sandrinecharles
> Hi @sandrinecharles, thanks for your submission to JOSS. So, it looks like the repository you submitted points to one for the paper, rather than the software itself. The typical approach we expect is for the paper to be included in the software repository itself, and then we just have a single repo (for the submitted software package).
>
> Can you move your paper to the repository for the software itself, and let us know when you have done that? Thanks!

Dear Kyle, I moved the 'paper' folder to the Github repository of the R-package directly as requested:
https://github.com/pveber/morse/tree/master/paper
