### 2021-04-20 #3177 @sandrinecharles
> Hello human, I'm @whedon, a robot that can help you with some common editorial tasks.
>
> **⚠️ JOSS reduced service mode ⚠️**
>
> Due to the challenges of the COVID-19 pandemic, JOSS is currently operating in a "reduced service mode". You can read more about what that means in [our blog post](https://blog.joss.theoj.org/2020/05/reopening-joss).
>
> For a list of things I can do to help you, just type:
>
> ```
> @whedon commands
> ```
>
> For example, to regenerate the paper pdf after making changes in the paper's md or bib files, type:
>
> ```
> @whedon generate pdf
> ```
