### 2021-04-21 #3177 @KristinaRiemer
Thanks for the suggestion for reviewers @sandrinecharles, I'm reaching out via email to those whose GitHub handles I couldn't find. Otherwise, @BrandonEdwards @puruckertom would you be interested and able to review this JOSS submission?
