### 2021-04-22 #3177 @sandrinecharles
> > Thanks for the suggestion for reviewers @sandrinecharles, I'm reaching out via email to those whose GitHub handles I couldn't find. Otherwise, @BrandonEdwards @puruckertom would you be interested and able to review this JOSS submission?
>
> Certainly interested, and certainly able to!

Thank you very much!
