### 2021-07-13 #3200 @sandrinecharles
Hi @KristinaRiemer thank you for taking care about this submission; it is a pity that the process takes a so long time to be achieved regarding the short length of the paper. I will be on holidays from July the 23rd, able to remain partially connected in case reviewers have questions. I really hope sometime happens soon.
Have a good holiday time, best regards, Sandrine
