### 2021-07-13 #3200 @KristinaRiemer
Hi @sandrinecharles, sorry about the delays. We do our best to make the JOSS review process fairly timely and straightforward, but it doesn't always work out that way unfortunately. Thanks for your patience!
