### 2021-08-21 #3200 @sandrinecharles
Dear @KristinaRiemer thanks a lot to follow the review process of our paper ; I cannot understand why it is so long, that is really a pity and could discourage us to submit other manuscripts to JOSS. What is happening with reviewers? The paper is rather short so that it should be quick to review... Bets regards, Sandrine Charles
