### 2021-08-23 #3200 @KristinaRiemer
Hi @sandrinecharles, sorry about the delay, I'm sure this is frustrating. The reviewers are not just going to be reviewing the paper, but also looking at the software itself, so that can be a substantial commitment of time.

@BrandonEdwards has let me know that he will be doing his review this week!
