### 2021-09-13 #3200 @KristinaRiemer
Hi @sandrinecharles. To update you on the current status of the reviewers, I have heard from both reviewers several times via email and they have said they will be able to finish their reviews. If they have not been able to finish the initial reviews by this Friday, I will have to look for new reviewers. Thank you so much for being patient with this process!
