### 2021-09-16 #3200 @KristinaRiemer
Hi @sandrinecharles, sorry for this taking so long. I do want to emphasize that this is an unusual circumstance. Most JOSS submissions get through the review process faster than this, which is also generally faster than manuscripts at most scientific journals.

@BrandonEdwards @puruckertom this is a friendly reminder to have your reviews in by tomorrow please!
