+++ paper/README.md
@@ -1,0 +1,2 @@
+# morse_paper
+the R-package 'morse' paper for publication in JOSS
