### 2021-10-09 #3200 @sandrinecharles
@KristinaRiemer I can propose to contact:
-  Peter Vermeiren (P.Vermeiren@science.ru.nl), post-doc in the Netherlands
- Another idea would be to contact Alessio IPPOLITTO from EFSA (Alessio.IPPOLITO@efsa.europa.eu)
- Marie TRIJAU (marie.trijau@ibacon.com) or Benoît GOUSSEN (Benoit.Goussen@ibacon.com), both from the Ibacon company, could also accept.
- else Stephan REICHENBERGER from Knell company (sreichenberger@knoell.com)
- or Andre GERGS from Bayer (andre.gergs@bayer.com)
