### 2021-10-10 #3200 @puruckertom
I think this is a very useful R package and a well-written manuscript. I have added issues to the morse_paper repo relevant to the checklist items that are part of JOSS’ peer review process. I commend the authors for their code contribution and the extent of the available documentation for learning how to use the package.

List of issues

-	https://github.com/sandrinecharles/morse_paper/issues/1
-	https://github.com/sandrinecharles/morse_paper/issues/2
-	https://github.com/sandrinecharles/morse_paper/issues/3
-	https://github.com/sandrinecharles/morse_paper/issues/4
-	https://github.com/sandrinecharles/morse_paper/issues/5
-	https://github.com/sandrinecharles/morse_paper/issues/6
-	https://github.com/sandrinecharles/morse_paper/issues/7
-	https://github.com/sandrinecharles/morse_paper/issues/8
