### 2021-11-10 #3200 @Peter-Vermeiren
**Software paper -> Summary / Statement of need**
The summary mentions _“full compliance with regulatory requirements”_. I would have expected a bit of detail about that here, such as a reference or examples of relevant regulations, or the specific requirements you are referring to. I am not an expert on regulatory affairs. Would this package e.g. work for USA regulations as well as EU regulations?
