### 2021-11-10 #3200 @Peter-Vermeiren
**Software paper -> quality of writing**
- There were some minor spelling errors in the paper, which I indicated with comments in the attached manuscript file.
- I also added 3 points of clarification (a reference, clarification on what "_advanced and innovative methods_" are, and a clarification of a figure legend)

[Submitted manuscript - version 09 Nov 2021.pdf](https://github.com/openjournals/joss-reviews/files/7514270/Submitted.manuscript.-.version.09.Nov.2021.pdf)
