### 2021-11-10 #3200 @Peter-Vermeiren
**Function survDataCheck:** It would be nice to have a more informative message when everything works fine (currently: "_no message_"). How about: "_data in correct format_". Messages for other datatypes are informative, thanks.  https://github.com/pveber/morse/blob/master/R/survDataCheck.R
