### 2021-11-10 #3200 @Peter-Vermeiren
**Help documentation for survFit function:** The documentation seems to repeat itself 2 to 3 times.
R/survFit.R https://github.com/pveber/morse/blob/master/R/survFit.R
Minor spelling issue: L.9: The function \code{survFit} return -> return**s** with an “s”
