### 2021-11-10 #3200 @Peter-Vermeiren
Thank you for the opportunity to review this manuscript and R package.
- I found the package very useful, with all functionalities that I would expect. The package is well documented so I could easily find how to adapt functions, as well as understand what the functions do and what their underlying theory is.
- The paper is clear, with an easy-to-follow example. (It is nice that a more extensive tutorial is also available).
- I only have very minor comments (see my previous 4 comments). The package and paper seems to have been tested and thought through quite extensively already.
