### 2021-11-10 #3200 @Peter-Vermeiren
p.s. I checked the relevant boxes in the review checklist. I did not check "_automated tests_" as this is not applicable as far as I can see.
