### 2021-12-04 #3200 @sandrinecharles
@Peter-Vermeiren @KristinaRiemer
> **Software paper -> Summary / Statement of need** The summary mentions _“full compliance with regulatory requirements”_. I would have expected a bit of detail about that here, such as a reference or examples of relevant regulations, or the specific requirements you are referring to. I am not an expert on regulatory affairs. Would this package e.g. work for USA regulations as well as EU regulations?

Fully agree with this suggestion; we added three references of different OECD guidance documents in our summary, and cited again the reference to EFSA PPR Panel 2018 within section "Survival analysis "(end of the first §). OECD references where of course added to the final list.
