### 2021-12-04 #3200 @sandrinecharles
@Peter-Vermeiren @KristinaRiemer

> **Software paper -> quality of writing**
>
> * There were some minor spelling errors in the paper, which I indicated with comments in the attached manuscript file.
> * I also added 3 points of clarification (a reference, clarification on what "_advanced and innovative methods_" are, and a clarification of a figure legend)
>
> [Submitted manuscript - version 09 Nov 2021.pdf](https://github.com/openjournals/joss-reviews/files/7514270/Submitted.manuscript.-.version.09.Nov.2021.pdf)

Thanks for pointing out spelling errors; we followed all your proposed changes, and also clarified where you suggested to do. Regarding spelling errors you pointed out in the message produced by 'morse' after the MFx calculations, I put an issue directly in the 'morse' package gitlab. This should be corrected with the next release but not changed immediately. I remove the grid in the background of Figure 5 (left panel).
