### 2021-12-04 #3200 @sandrinecharles
@KristinaRiemer

> **Function survDataCheck:** It would be nice to have a more informative message when everything works fine (currently: "_no message_"). How about: "_data in correct format_". Messages for other datatypes are informative, thanks. https://github.com/pveber/morse/blob/master/R/survDataCheck.R

Thanks @Peter-Vermeiren , nice suggestion that should also apply to the reproDataCheck function in 'morse'... I put another issue in our 'morse' gitlab for the next release. I asked that the next release is made asap, hopefully just before this manuscript is accepted, if it is finally.
