### 2021-12-04 #3200 @sandrinecharles
@KristinaRiemer

> **Help documentation for survFit function:** The documentation seems to repeat itself 2 to 3 times.
> R/survFit.R https://github.com/pveber/morse/blob/master/R/survFit.R
> Minor spelling issue: L.9: The function \code{survFit} return -> return**s** with an “s”

OK @Peter-Vermeiren , again an issue in the 'morse' gitlab for the next release. Thanks a lot for all your comments.
