### 2021-12-04 #3200 @sandrinecharles
> p.s. I checked the relevant boxes in the review checklist.

@Peter-Vermeiren you are right, automated test are not available.
