+++ paper/paper.md
@@ -10,7 +10,7 @@ tags:
 - Bayesian inference
 - Survival
 - Reproduction
-date: "2021-11-05"
+date: "2021-04-07"
 output:
   pdf_document:
     extra_dependencies: ["float"]
@@ -29,7 +29,7 @@ bibliography: paper.bib
 ---
 
 # Summary
-Package `morse` is devoted to the analysis of experimental data collected from standard toxicity tests. It provides ready-to-use functions to visualize a data set and to estimate several toxicity indices to be further used in support of environmental risk assessment in full compliance with regulatory requirements. Such toxicity indices are indeed typically requested by standardized regulatory guidelines on which national agencies base their evaluation of applications for regulatory approval of chemical substances. Tools gathered together within package `morse` involve the most advanced and innovative methods developed in research, making them available for ecotoxicologists and regulators who so not need to invest in the underlying technicalities in order to perform a valuable quantitative environmental risk assessment.
+Package `morse` is devoted to the analysis of experimental data collected from standard toxicity tests. It provides ready-to-use functions to visualize a data set and to estimate several toxicity indices to be further used in support of environmental risk assessment in full compliance with regulatory requirements [@OECD2006]. Such toxicity indices are indeed typically requested by standardized regulatory guidelines on which national agencies base their evaluation of applications for regulatory approval of chemical substances; see  for example OECD toxicity testing guideline [@OECD2016; @OECD2018]. Tools gathered together within package `morse` involve the most advanced and innovative methods developed for ecotoxicology, such as appropriate stochastic parts in dose-response modelling [@MLDM2014], or the use of Bayesian statistics to get parameter estimates as posterior probability distribution [@Billoir2008]. Hence, these tools are easily accessible for ecotoxicologists and regulators who do not need to deeply invest in the underlying technicalities in order to perform a valuable quantitative environmental risk assessment.
 
 This paper provides an overview of a typical use of package `morse` with survival data collected over time and at different increasing exposure concentrations. These data are analysed with the reduced version of GUTS models based on the stochastic death hypothesis (namely, the `GUTS-RED-SD` model). This example can be followed step-by-step to analyse any new data set, as long as the data set format is respected.
 
@@ -40,7 +40,7 @@ Package `morse` [@morse2021] has been tested using `R` (version 3.5 and later) o
 
 All functions in package `morse` can be used without a deep knowledge of their underlying probabilistic model or inference methods. Rather, they were designed to behave as well as possible, without requiring the user to provide values for some parameters. Nevertheless, models implemented in `morse` can also be used as a first step to tailor new models for more specific situations.
 
-Note that package `morse` benefits from a web interface, MOSAIC, from which the same analyses can be reproduced directly on-line without having to implement in `R` programming. MOSAIC is freely available at [https://mosaic.univ-lyon1.fr/](https://mosaic.univ-lyon1.fr/) [@Charles2017] (Figure \ref{fig:mosaic}).
+Note that package `morse` benefits from a web interface, MOSAIC, from which the same analyses can be reproduced directly on-line without having to implement them directly in `R` programming. MOSAIC is freely available at [https://mosaic.univ-lyon1.fr/](https://mosaic.univ-lyon1.fr/) [@Charles2017] (Figure \ref{fig:mosaic}).
 
[…]
