+++ paper/paper.bib
@@ -1,3 +1,64 @@
+@article{Billoir2008,
+author = {Billoir, Elise and Delignette-Muller, Marie Laure and P{\'{e}}ry, Alexandre R R and Charles, Sandrine and Pery, A R R and Charles, Sandrine},
+file = {:Users/scharles/Documents/Mendeley Desktop/2008/Billoir et al/Billoir et al. - 2008 - Environmental Science {\&} Technology(2).pdf:pdf;:Users/scharles/Documents/Mendeley Desktop/2008/Billoir et al/Billoir et al. - 2008 - Environmental Science {\&} Technology.pdf:pdf},
+issn = {0013-936X},
+journal = {Environmental Science {\&} Technology},
+keywords = {Animals,Bayes Theorem,Biological,Copper,Copper: toxicity,Daphnia,Daphnia: drug effects,Daphnia: physiology,Databases as Topic,Ecotoxicology,Models,Reproduction,Reproduction: drug effects,Zinc,Zinc: toxicity},
+month = {dec},
+number = {23},
+pages = {8978--84},
+pmid = {19192828},
+title = {{A Bayesian Approach to Analyzing Ecotoxicological Data}},
+url = {http://www.ncbi.nlm.nih.gov/pubmed/19192828},
+volume = {42},
+year = {2008}
+}
+
+
+@article{MLDM2014,
+author = {Delignette-Muller, Marie Laure and Lopes, Christelle and Veber, Philippe and Charles, Sandrine},
+doi = {10.1021/es502009r},
+file = {:Users/scharles/Documents/Mendeley Desktop/2014/Delignette-Muller et al/Delignette-Muller et al. - 2014 - Environmental Science {\&} Technology.pdf:pdf},
+issn = {1520-5851},
+journal = {Environmental Science {\&} Technology},
+month = {jul},
+number = {13},
+pages = {7544--51},
+pmid = {24892187},
+title = {{Statistical handling of reproduction data for exposure-response modeling.}},
+url = {http://www.ncbi.nlm.nih.gov/pubmed/24892187},
+volume = {48},
+year = {2014}
+}
+
+
+@book{OECD2006,
+   author = "OCDE",
+   title = "Current Approaches in the Statistical Analysis of Ecotoxicity Data",
+   year = "2006",
+   pages = 147,
+   url = "https://www.oecd-ilibrary.org/content/publication/9789264085275-en",
+   doi = "https://doi.org/10.1787/9789264085275-en" 
+}
+
+@book{OECD2016,
+   author = "OCDE",
+   title = "Test No. 243: Lymnaea stagnalis Reproduction Test",
+   year = "2016",
+   pages = 31,
+   url = "https://www.oecd-ilibrary.org/content/publication/9789264264335-en",
+   doi = "https://doi.org/10.1787/9789264264335-en" 
+}
+
+@book{OECD2018,
+   author = "OCDE",
+   title = "Daphnia magna Reproduction Test (OECD TG 211)",
+   year = "2018",
+   pages = ,
+   url = "https://www.oecd-ilibrary.org/content/component/9789264304741-12-en",
+   doi = "https://doi.org/10.1787/9789264304741-12-en" 
+}
+
 @article{Charles2017,
 author = {Charles, Sandrine and Delignette-Muller, Marie Laure and Veber, Philippe and Delignette-Muller, Marie Laure},
 doi = {10.1007/s11356-017-9809-4},
