### 2021-12-04 #3200 @sandrinecharles
> Thank you for the opportunity to review this manuscript and R package.
>
> * I found the package very useful, with all functionalities that I would expect. The package is well documented so I could easily find how to adapt functions, as well as understand what the functions do and what their underlying theory is.
> * The paper is clear, with an easy-to-follow example. (It is nice that a more extensive tutorial is also available).
> * I only have very minor comments (see my previous 4 comments). The package and paper seems to have been tested and thought through quite extensively already.

@Peter-Vermeiren Thanks again for your helpful comments leading to improvement in our paper but also in the 'morse' package.
@KristinaRiemer Thank you for having managed the review process. I achieved all changes related to comments from @Peter-Vermeiren. Accordingly, the gitlab dossier with paper files has been fully updated: https://github.com/pveber/morse/tree/master/paper.

Best regards,
Sandrine
