### 2021-12-06 5d3e806
MFx plot function: spelling in message #289
