+++ R/plot.MFx.R
@@ -125,8 +125,8 @@ plot.MFx <- function(x,
                    aes(x = x.pts, y = y.pts, color = pts.leg)) 
       
       warning("This is not an error message:
-Just take into account that MFx as been estimated with a binary
-search using the 'accuracy' argument. Cross point indicate the
+Just take into account that MFx has been estimated with a binary
+search using the 'accuracy' argument. Cross points indicate the
 position of evaluated time series. To improve the shape of the curve, you 
 can use X = NULL, and compute time series around the median MFx, with the
           vector `MFx_range`.")
