### 2021-12-06 a4378b7
message data checking: change 'No message' to 'Correct message' #290
