### 2021-12-06 b3ef6c7
survFit documentation: fix rdname to stop repeating documentation #291
