### 2021-12-08 #3200 @sandrinecharles
@KristinaRiemer According to reviewer's comments, we submitted the new release of our 'morse' package to the official CRAN web site two days ago.
