### 2021-12-08 #3200 @KristinaRiemer
These are the handful of typos/formatting that I was able to find in the paper:

- consistency in capitalization for TKTD full length (lines 29 & 37)
- "team in performing" to "team to perform"? (line 39-40)
- "and" between EASYGUTS and GATEAUX in model list (line 41)
- "can be either" to "can either be" (line 66)
- "the Survival probability prediction error at the end of the exposure profile (SPPE)" to "the Survival Probability Prediction Error (SPPE) at the end of the exposure profile" or just capitalize "Survival Probability Prediction Error"? (lines 82-83)

I will wait to check the references for formatting, DOIs, etc. when the whedon error is resolved.
