### 2021-12-09 #3200 @sandrinecharles
>

> > Checking the BibTeX entries failed with the following error:
> > Failed to parse BibTeX on value "," (COMMA) [#<BibTeX::Bibliography data=[4]>, "@", #<BibTeX::Entry >, {:author=>["OCDE"], :title=>["Daphnia magna Reproduction Test (OECD TG 211)"], :year=>["2018"]}, ",", "pages", "="]
>
> I think this whedon reference check error is occurring because the "Daphnia magna Reproduction Test" reference needs a number in the pages entry.

OK I will check; this is probably due to the fact that this reference in a regulatory document not a standard article.
