+++ paper/paper.bib
@@ -37,8 +37,7 @@ year = {2014}
    title = "Current Approaches in the Statistical Analysis of Ecotoxicity Data",
    year = "2006",
    pages = 147,
-   url = "https://www.oecd-ilibrary.org/content/publication/9789264085275-en",
-   doi = "https://doi.org/10.1787/9789264085275-en" 
+   doi = "10.1787/9789264085275-en" 
 }
 
 @book{OECD2016,
@@ -46,17 +45,15 @@ year = {2014}
    title = "Test No. 243: Lymnaea stagnalis Reproduction Test",
    year = "2016",
    pages = 31,
-   url = "https://www.oecd-ilibrary.org/content/publication/9789264264335-en",
-   doi = "https://doi.org/10.1787/9789264264335-en" 
+   doi = "10.1787/9789264264335-en" 
 }
 
-@book{OECD2018,
+@book{OECD2012,
    author = "OCDE",
    title = "Daphnia magna Reproduction Test (OECD TG 211)",
-   year = "2018",
-   pages = ,
-   url = "https://www.oecd-ilibrary.org/content/component/9789264304741-12-en",
-   doi = "https://doi.org/10.1787/9789264304741-12-en" 
+   year = "2012",
+   pages = 25,
+   doi = "10.1787/9789264304741-12-en" 
 }
 
 @article{Charles2017,
