### 2021-12-09 88e9444
text updates
Some typos updates.
