### 2021-12-09 #3200 @sandrinecharles
> * consistency in capitalization for TKTD full length (lines 29 & 37)

Caps are now removed on line 29, while the abbreviation TKTD is directly used on line 37.

> * "team in performing" to "team to perform"? (line 39-40)

done

> * "and" between EASYGUTS and GATEAUX in model list (line 41)

done

> * "can be either" to "can either be" (line 66)

done

> * "the Survival probability prediction error at the end of the exposure profile (SPPE)" to "the Survival Probability Prediction Error (SPPE) at the end of the exposure profile" or just capitalize "Survival Probability Prediction Error"? (lines 82-83)

Changes as "and the Survival Probability Prediction Error (SPPE) at the end of the exposure profile.".

Additionally, line 42, "frequestist" was changed to "frequentist".

The github repository of the paper has been updated accordingly.
