### 2021-12-09 #3200 @KristinaRiemer
Thanks for making those edits to the text!

The references look pretty good, these were the only small errors I found:

- replace link in ref for "A Bayesian Approach to Analyzing Ecotoxicological Data" with DOI (see whedon reference check above)
- that same reference also has you listed as author twice
- the "Modelling survival under chemical stress" reference link returned that that page isn't found; is the ISBN number right?
- For last three references, should "OCDE" be "OECD"?
