+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{Billoir2008,
-author = {Billoir, Elise and Delignette-Muller, Marie Laure and P{\'{e}}ry, Alexandre R R and Charles, Sandrine and Pery, A R R and Charles, Sandrine},
+author = {Billoir, Elise and Delignette-Muller, Marie Laure and Pery, Alexandre R R and Charles, Sandrine},
 file = {:Users/scharles/Documents/Mendeley Desktop/2008/Billoir et al/Billoir et al. - 2008 - Environmental Science {\&} Technology(2).pdf:pdf;:Users/scharles/Documents/Mendeley Desktop/2008/Billoir et al/Billoir et al. - 2008 - Environmental Science {\&} Technology.pdf:pdf},
 issn = {0013-936X},
 journal = {Environmental Science {\&} Technology},
@@ -7,9 +7,8 @@ keywords = {Animals,Bayes Theorem,Biological,Copper,Copper: toxicity,Daphnia,Dap
 month = {dec},
 number = {23},
 pages = {8978--84},
-pmid = {19192828},
+doi = {10.1021/es801418x},
 title = {{A Bayesian Approach to Analyzing Ecotoxicological Data}},
-url = {http://www.ncbi.nlm.nih.gov/pubmed/19192828},
 volume = {42},
 year = {2008}
 }
@@ -33,7 +32,7 @@ year = {2014}
 
 
 @book{OECD2006,
-   author = "OCDE",
+   author = "OECD",
    title = "Current Approaches in the Statistical Analysis of Ecotoxicity Data",
    year = "2006",
    pages = 147,
@@ -41,7 +40,7 @@ year = {2014}
 }
 
 @book{OECD2016,
-   author = "OCDE",
+   author = "OECD",
    title = "Test No. 243: Lymnaea stagnalis Reproduction Test",
    year = "2016",
    pages = 31,
@@ -49,7 +48,7 @@ year = {2014}
 }
 
 @book{OECD2012,
-   author = "OCDE",
+   author = "OECD",
    title = "Daphnia magna Reproduction Test (OECD TG 211)",
    year = "2012",
    pages = 25,
@@ -110,8 +109,6 @@ year = {2019}
 @book{Jager2018,
 author = {Jager, Tjalling and Ashauer, Roman},
 edition = {Leanpub},
-file = {:Users/scharles/Documents/Mendeley Desktop/2018/Jager, Ashauer/Jager, Ashauer - 2018 - Unknown(2).pdf:pdf},
-isbn = {9781999970505},
 pages = {Version 1.0},
 title = {{Modelling survival under chemical stress. A comprehensive guide to the GUTS framework}},
 url = {https://leanpub.com/guts_book},
@@ -121,8 +118,6 @@ year = {2018}
 @article{EFSA2018,
 author = {{EFSA PPR Panel}},
 doi = {10.2903/j.efsa.2018.5377},
-file = {:Users/scharles/Documents/Mendeley Desktop/2018/EFSA PPR Panel/EFSA PPR Panel - 2018 - EFSA Journal.pdf:pdf},
-issn = {18314732},
 journal = {EFSA Journal},
 keywords = {aquatic organisms,model calibration,model evaluation,model validation,prospective risk assessment,time-variable exposure,toxicodynamic models,toxicokinetic},
 number = {8},
