### 2021-12-09 #3200 @sandrinecharles
> * replace link in ref for "A Bayesian Approach to Analyzing Ecotoxicological Data" with DOI (see whedon reference check above)

DOI added

> * that same reference also has you listed as author twice

corrected

> * the "Modelling survival under chemical stress" reference link returned that that page isn't found; is the ISBN number right?

I removed the isbn number as it does not appear ont the web page. The URL is correct: https://leanpub.com/guts_book.

> * For last three references, should "OCDE" be "OECD"?

Yes! changed.

The github repository of the paper has been updated accordingly just few minutes ago.
