### 2021-12-09 #3200 @KristinaRiemer
Great, thanks for making all those changes! The paper looks good now to me.

The next steps are to: 1) [create a tagged release](https://docs.github.com/en/repositories/releasing-projects-on-github/managing-releases-in-a-repository#creating-a-release) on GitHub and 2) archive the code (on Zenodo, figshare, institutional repository, etc.) to generate a DOI. If you can post a link to the DOI here when you're done please.
