### 2021-12-10 #3200 @danielskatz
@sandrinecharles - I've suggested a bunch of changes in https://github.com/pveber/morse/pull/292 - please merge this (or let me know what you disagree with), then run `@whedon generate pdf` to make sure everything still looks good to you, and let me know when you are ready to proceed to publication
