### 2021-12-10 #3200 @sandrinecharles
@danielskatz A last edit change is needed at line 127:
"We will used the reduced version" --> "We will use the reduced version"
