### 2021-12-10 #3200 @sandrinecharles
> Please go ahead and make that change, then let me know

done, in push to https://github.com/pveber/morse/tree/master/paper
