### 2021-12-10 #3200 @danielskatz
@sandrinecharles - you can go back to https://github.com/pveber/morse/blob/a5ed566cd4d9849b2105bd9b05c757a6d2f92be1/paper/paper.md and make the appropriate edit there
