### 2021-12-11 #3200 @sandrinecharles
@KristinaRiemer Dear Kristina, we juste finalised a new paper that we would also like to submit to JOSS, on a new package developed this year: https://CRAN.R-project.org/package=rbioacc, in support of a dedicated Shiny tool available fro free online: http://lbbe-shiny.univ-lyon1.fr/mosaic-bioacc/. Would you be interested in taking care of the reviewing process for this paper. Peter Vermeiren could be one the reviewer, as the philosophy is really the same as in 'morse'. I could also suggest to ask Benoît Goussen from Section Head Ecological Modelling, ibacon GmbH, in Germany. Our preprint is already within bioRxiv here: https://www.biorxiv.org/content/10.1101/2021.09.08.459421v2, and our Git repository here: https://gitlab.in2p3.fr/sandrine.charles/rbioacc
