### 2022-01-05 #3200 @kthyng
@sandrinecharles In case you haven't already done so, please just go ahead and submit your new JOSS paper and you can select @KristinaRiemer as the suggested editor. Please include your reviewer suggestion.
